+++ examples/pyplot/fitPolynomial2.py
@@ -16,7 +16,7 @@ yerrs = np.linspace(1.0, 0.4, n)  # make first points less precise
 noise = np.random.randn(n)
 
 # Plot the noisy points with their error bars
-plt = plot(x, y+noise, '.k',
+fig = plot(x, y+noise, '.k',
            title=__doc__+str(deg),
            xerrors=xerrs,
            yerrors=yerrs,
@@ -24,7 +24,7 @@ plt = plot(x, y+noise, '.k',
            xlim=(-3,15),
            ylim=(-3,15),
           )
-plt += DashedLine(x, y)
+fig += DashedLine(x, y)
 
 # Fit points and evaluate, with a boostrap and Monte-Carlo technique,
 # the correct errors and error bands. Return a Line object:
@@ -37,12 +37,12 @@ pfit = fit([x, y+noise],
            vrange=(-3,15), # specify the domain of fit
           )
 
-plt += [pfit, pfit.errorBand, *pfit.errorLines] # add these objects to Plot
+fig += [pfit, pfit.errorBand, *pfit.errorLines] # add these objects to Plot
 
 txt = "fit coefficients:\n " + precision(pfit.coefficients, 2) \
     + "\n\pm" + precision(pfit.coefficientErrors, 2) \
     + "\n\Chi^2_\nu  = " + precision(pfit.reducedChi2, 3)
-plt += Text3D(txt, s=0.42, font='VictorMono').pos(2,-2).c('k')
+fig += Text3D(txt, s=0.42, font='VictorMono').pos(2,-2).c('k')
 
 # Create an histo to show the correlation of fit parameters
 h = histogram(pfit.MonteCarloCoefficients[:,0],
@@ -52,4 +52,4 @@ h = histogram(pfit.MonteCarloCoefficients[:,0],
               cmap='bone_r', scalarbar=True)
 h.scale(150).shift(-1,11) # make it a lot bigger and move it
 
-show(plt, h, zoom=1.3, mode="image").close()
+show(fig, h, zoom=1.3, mode="image").close()