+++ examples/pyplot/histo_hexagonal.py
@@ -7,15 +7,16 @@ x = np.random.randn(N) * 1.0
 y = np.random.randn(N) * 1.5
 
 # hexagonal binned histogram:
-histo = histogram(x, y,
-                  bins=10,
-                  mode='hexbin',
-                  xtitle="\sigma_x =1.0",
-                  ytitle="\sigma_y =1.5",
-                  ztitle="counts",
-                  fill=True,
-                  cmap='terrain',
-                  )
+histo = histogram(
+    x, y,
+    bins=10,
+    mode='hexbin',
+    xtitle="\sigma_x =1.0",
+    ytitle="\sigma_y =1.5",
+    ztitle="counts",
+    fill=True,
+    cmap='terrain',
+)
 # add a formula:
 f = r'f(x, y)=A \exp \left(-\left(\frac{\left(x-x_{o}\right)^{2}}'
 f+= r'{2 \sigma_{x}^{2}}+\frac{\left(y-y_{o}\right)^{2}}'