+++ examples/pyplot/plot_empty.py
@@ -1,7 +1,7 @@
-"""Create an empty Plot to be filled in a loop
-Any 3D Mesh object can be added to the plot as well"""
-from vedo import *
-from vedo.pyplot import plot, Plot
+"""Create an empty Figure to be filled in a loop
+Any 3D Mesh object can be added to the figure as well"""
+from vedo import settings, dataurl, Mesh, Circle, show
+from vedo.pyplot import plot, Figure
 import numpy as np
 
 
@@ -12,19 +12,19 @@ x = np.linspace(0, 4*np.pi, 200)
 y = np.sin(x) * np.sin(x/12)
 
 # dictionary of options for the axes
-ax_opts = dict(xtitle="distance", xyGrid=False, axesLineWidth=3, xyFrameLine=3)
+ax_opts = dict(xtitle="distance", xyGrid=1, axesLineWidth=3, xyFrameLine=3)
 
 # Create an empty Plot and fill it
-pl = Plot(xlim=(0, 12), ylim=(-1.2, 1.2), aspect=16/9, axes=ax_opts)
+fig = Figure(xlim=(0,12), ylim=(-1.1, 1.1), aspect=16/9, axes=ax_opts)
 for i in range(10):
-    pl += plot(x, y * i/5, lc=i)  # lc= line color (10-colors palette index)
+    fig += plot(x, y * i/5, lc=i)  # lc= line color (10-colors palette index)
 
 # Add any number of polygonal Meshes.
 # Use add() to keep the object aspect ratio inside the Plot coord system:
-mesh = Mesh(dataurl+'cessna.vtk').c('green4').scale(0.4)
-circle = Circle([5,0.5, 1], c='orange5')
-pl.add(mesh.pos(4, 0.5, 2), circle)
+mesh = Mesh(dataurl+'cessna.vtk').c('blue5').scale(0.4).pos(4, 0.5)
+circle = Circle([5,0.5], c='orange5')
+fig.add(mesh, circle, as3d=True)
 
-show(pl, __doc__, size=(1000,800), zoom='tight').close()
+show(fig, __doc__, size=(1000,700), zoom=1.5).close()